requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
tensorflow-cpu==2.16.2
pillow==10.4.0
numpy==1.26.4
python-multipart==0.0.12
